move files and make separate template files
