+++ apps/browser/src/platform/popup/layout/popup-footer.component.html
@@ -1,0 +1,7 @@
+<footer class="tw-p-3 tw-border-0 tw-border-solid tw-border-t tw-border-secondary-300">
+  <div class="tw-max-w-screen-sm tw-mx-auto">
+    <div class="tw-flex tw-justify-start">
+      <ng-content select="[actionFooter]"></ng-content>
+    </div>
+  </div>
+</footer>
