+++ apps/browser/src/platform/popup/layout/popup-layout.component.html
@@ -1,0 +1,9 @@
+<div class="tw-border tw-h-full tw-border-secondary-300 tw-border-solid tw-flex tw-flex-col">
+  <ng-content select="[popupHeader]"></ng-content>
+  <main class="tw-bg-background-alt tw-p-3 tw-flex-1 tw-overflow-y-scroll">
+    <div class="tw-max-w-screen-sm tw-mx-auto">
+      <ng-content></ng-content>
+    </div>
+  </main>
+  <ng-content select="[popupFooter]"></ng-content>
+</div>
