+++ apps/browser/src/platform/popup/layout/popup-layout.component.ts
@@ -1,0 +1,8 @@
+import { Component } from "@angular/core";
+
+@Component({
+  selector: "popup-layout",
+  templateUrl: "popup-layout.component.html",
+  standalone: true,
+})
+export class PopupLayoutComponent {}
