+++ libs/components/tailwind.config.js
@@ -5,6 +5,7 @@ config.content = [
   "libs/components/src/**/*.{html,ts,mdx}",
   "libs/auth/src/**/*.{html,ts,mdx}",
   "apps/web/src/**/*.{html,ts,mdx}",
+  "apps/browser/src/**/*.{html,ts,mdx}",
   "bitwarden_license/bit-web/src/**/*.{html,ts,mdx}",
   ".storybook/preview.tsx",
 ];
